+++ R/extract_folds_colnames.R
@@ -1,0 +1,5 @@
+
+
+extract_fold_colnames <- function(data, regex="^.folds"){
+  colnames(data)[grep(regex, colnames(data))]
+}